@@ -13,19 +13,19 @@
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.24",
     "@types/node": "^12.20.37",
-    "@types/react": "^17.0.37",
     "@types/react-bootstrap": "^0.32.28",
-    "@types/react-dom": "^17.0.11",
     "aws-amplify": "^4.3.10",
     "axios": "^0.21.4",
     "bootstrap": "^5.1.3",
     "is-descriptor": "^3.0.0",
+    "leaflet": "^1.8.0",
     "prettier": "^2.5.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-apexcharts": "^1.4.0",
     "react-bootstrap": "^2.0.3",
     "react-charts": "^3.0.0-beta.47",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
+    "react-leaflet": "^4.0.1",
     "react-router-dom": "^6.1.1",
     "react-scripts": "4.0.3",
     "react-simple-maps": "^2.3.0",
@@ -47,20 +47,16 @@
       "react-app/jest"
     ]
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not op_mini all"
+  ],
   "devDependencies": {
+    "@types/leaflet": "^1.7.11",
     "@types/lodash": "^4.14.182",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
     "@types/react-simple-maps": "^1.0.7"
   }
 }
