@@ -17,28 +17,34 @@ export const AuthContext = createContext<AuthContextType>(null!);
 export const AuthProvider = ({ children }: { children: ReactNode }) => {
   const [user, setUser] = useState<IUser | null>(null);
 
+  const createUser = async (): Promise<IUser | null> => {
+    const user = await getCurrentUser();
+    console.log(user);
+    if (!user) {
+      return null;
+    }
+    return {
+      firstName: user.attributes.given_name,
+      lastName: user.attributes.family_name,
+      userId: user.attributes.sub,
+      role: user.attributes["custom:role"],
+      email: user.attributes.email,
+    };
+  };
+
   useEffect(() => {
     const isLoggedIn = async () => {
-      const user = await getCurrentUser();
-      if (user) {
-        const newUser = {
-          role: user.attributes["custom:role"],
-          email: user.attributes.email,
-        };
-        setUser(newUser);
-      }
+      const user = await createUser();
+      setUser(user);
     };
     isLoggedIn();
   }, []);
 
   const signIn = async (email: string, password: string) => {
     await cognitoSignIn(email, password);
-    const cognitoUser = await getCurrentUser();
-    const newUser = {
-      role: cognitoUser.attributes["custom:role"],
-      email: cognitoUser.attributes.email,
-    };
-    setUser(newUser);
+
+    const user = await createUser();
+    setUser(user);
   };
 
   const signOut = async () => {
