@@ -10,8 +10,8 @@ import { ISolarPanel } from "../Models/ISolarPanel";
 
 interface IProps {
   allSolarPanels: ISolarPanel[];
-  solarPanelType: ISolarPanel | null;
-  setSolarPanelType: (solarPanel: ISolarPanel) => void;
+  selectedSolarPanel: string;
+  setSelectedSolarPanel: (solarPanel: string) => void;
 }
 
 export const SolarPanelType = (props: IProps) => {
@@ -23,24 +23,32 @@ export const SolarPanelType = (props: IProps) => {
     ));
   };
 
-  const handleChange = (solarPanelName: string) => {
-    const chosenPanel = props.allSolarPanels.find(
-      (solarPanel) => solarPanel.name === solarPanelName
+  const getSolarPanelPowerAndPrice = (): { power: number; price: number } => {
+    const selectedPanel = props.allSolarPanels.find(
+      (panel) => panel.name === props.selectedSolarPanel
     );
-    if (chosenPanel) {
-      props.setSolarPanelType(chosenPanel);
+
+    if (!selectedPanel) {
+      return {
+        power: 0,
+        price: 0,
+      };
     }
+
+    return {
+      power: selectedPanel.power,
+      price: selectedPanel.price,
+    };
   };
 
   return (
     <Card
       style={{
         padding: "15px",
+        height: "100%",
       }}
     >
-      <Typography sx={{ textAlign: "center" }}>
-        CHOOSE SOLAR PANEL TYPE
-      </Typography>
+      <Typography sx={{ textAlign: "center" }}>SOLAR PANELS</Typography>
       <Divider />
       <Box
         sx={{
@@ -48,14 +56,33 @@ export const SolarPanelType = (props: IProps) => {
           textAlign: "center",
         }}
       >
-        <Typography>Solar Panel Type:</Typography>
+        <Typography>Type:</Typography>
         <Select
-          value={props.solarPanelType?.name}
+          value={props.selectedSolarPanel}
           sx={{ width: "100%", mt: "10px", borderTop: "1px solid grey" }}
-          onChange={(e) => handleChange(e.target.value as string)}
+          onChange={(e) =>
+            props.setSelectedSolarPanel(e.target.value as string)
+          }
         >
           {getMenuItems()}
         </Select>
+      </Box>
+      <Box
+        sx={{
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "center",
+          mt: "20px",
+        }}
+      >
+        <Typography sx={{ mr: "10px" }}>
+          {" "}
+          Power: <b>{getSolarPanelPowerAndPrice().power}W</b>{" "}
+        </Typography>
+        <Typography>
+          {" "}
+          Price: <b>{getSolarPanelPowerAndPrice().price}$</b>{" "}
+        </Typography>
       </Box>
       <Typography sx={{ textAlign: "center", marginTop: "10px" }}>
         Choose the specific solar panel type regarding of power and price.
