@@ -1,12 +1,17 @@
 import { ISolarPanel } from "./ISolarPanel";
 import { IEnvironmentalImpact } from "./IEnvironmentalImpact";
+import { ICity } from "./ICity";
+import { IUser } from "./IUser";
 
 export interface IInvestment {
   _id?: string;
-  userId?: string;
-  city: string;
+  user: IUser;
+  city: ICity;
   date: Date;
-  roofSize: number;
+  numberOfPanels: number;
+  power: number;
+  cost: number;
+  moneySaved: number;
   solarPanel: ISolarPanel;
   environmentalImpact: IEnvironmentalImpact;
   monthlyBillPrice: number;
